+++ HISTORY.rst
@@ -19,8 +19,8 @@ Old name                           New name
 ``github3.iter_all_repos``         ``github3.all_repos``
 ``github3.iter_all_users``         ``github3.all_users``
 ``github3.iter_events``            ``github3.all_events``
-``github3.iter_followers``         ``github3.followers``
-``github3.iter_following``         ``github3.following``
+``github3.iter_followers``         ``github3.followers_of``
+``github3.iter_following``         ``github3.followed_by``
 ``github3.iter_repo_issues``       ``github3.repo_issues``
 ``github3.iter_orgs``              ``github3.organizations``
 ``github3.iter_user_repos``        ``github3.user_repos``